refactor(sema): Split 6,611-line monolith into focused modules
Split crates/rue-air/src/sema.rs into a sema/ directory with 8 modules:
- mod.rs (398 lines): Public API, Sema struct, entry points
- context.rs (365 lines): Internal helper types (LocalVar, AnalysisContext, etc.)
- declarations.rs (718 lines): Declaration gathering, type registration
- builtins.rs (143 lines): Built-in type injection (String, etc.)
- typeck.rs (288 lines): Type resolution and checking helpers
- analysis.rs (4,422 lines): Core function body analysis
- tests.rs (185 lines): Unit tests
- airgen.rs (4 lines): Stub for future expansion

All 146 unit tests pass. All 1172 relevant spec tests pass.

Closes rue-rxws



resolve conflict: apply Copy array fix to refactored sema module
