+++ crates/rue-air/src/sema/airgen.rs
@@ -1,0 +1,4 @@
+//! AIR generation helpers.
+//!
+//! This module is kept for future expansion but currently most AIR generation
+//! is handled directly in the analysis module.
